SDK-1659: Move endpoint generation functions to live in their respective packages
Will rename dynamic_sharing_service in a separate PR
